Fix gnss_crpa_alpha console output: this project's console_print has no %u/%lu
Confirmed on real hardware: gnss_crpa_alpha=1.0 printed "raw=" with
nothing after it. console.c's console_print() is a hand-rolled minimal
formatter (not real printf) whose switch only handles %c/%s/%d/%x/%f --
no %u case exists, so %u silently consumes nothing and prints nothing;
%lu is worse, since the unhandled 'l' falls through and 'u' gets
printed as a literal character, garbling the surrounding text too.

set_gnss_crpa_alpha() and get_gnss_crpa_alpha() used %u for the raw
alpha value and %lu for the version-mismatch warning. Replaced both
with %d (long), matching how every other command in this file already
prints values through console_print (e.g. get_gnss_tx's (long)pass_en
pattern) -- %d already reads a long via va_arg here, so casting to
(long) instead of (unsigned)/(unsigned long) prints identically for
these non-negative 16-bit values.

+++ firmware/app_gnss_e310/command.c
@@ -403,9 +403,9 @@ void set_gnss_crpa_alpha(double* param, char param_no)
 	}
 
 	gnss_pt_set_crpa_alpha(param[0]);
-	console_print("GNSS_CRPA_ALPHA: set to %.4f (raw=%u)\n",
+	console_print("GNSS_CRPA_ALPHA: set to %.4f (raw=%d)\n",
 		      gnss_pt_get_crpa_alpha(),
-		      (unsigned)gnss_pt_get_crpa_alpha_raw());
+		      (long)gnss_pt_get_crpa_alpha_raw());
 }
 
 /**************************************************************************//***
@@ -416,17 +416,17 @@ void get_gnss_crpa_alpha(double* param, char param_no)
 {
 	uint32_t version = gnss_pt_read(GNSS_PT_REG_VERSION);
 
-	console_print("GNSS_CRPA_ALPHA: %.4f (raw=%u)\n",
+	console_print("GNSS_CRPA_ALPHA: %.4f (raw=%d)\n",
 		      gnss_pt_get_crpa_alpha(),
-		      (unsigned)gnss_pt_get_crpa_alpha_raw());
+		      (long)gnss_pt_get_crpa_alpha_raw());
 	if(version != GNSS_PT_EXPECTED_VERSION) {
 		console_print("                 WARNING: bitstream reports version "
-			      "%lu.%lu, not %lu.%lu -- the CRPA core may not be "
+			      "%d.%d, not %d.%d -- the CRPA core may not be "
 			      "present, so this value may not be consumed by "
 			      "anything.\n",
-			      (unsigned long)(version >> 16), (unsigned long)(version & 0xFFFFU),
-			      (unsigned long)(GNSS_PT_EXPECTED_VERSION >> 16),
-			      (unsigned long)(GNSS_PT_EXPECTED_VERSION & 0xFFFFU));
+			      (long)(version >> 16), (long)(version & 0xFFFFU),
+			      (long)(GNSS_PT_EXPECTED_VERSION >> 16),
+			      (long)(GNSS_PT_EXPECTED_VERSION & 0xFFFFU));
 	}
 }
